# Add rayon feature with optimized two-pass resample with rayon enabled