--- a/Cargo.toml
+++ b/Cargo.toml
@@ -14,11 +14,15 @@ include = ["Cargo.toml", "README.md", "LICENSE", "src/*.rs"]
 edition = "2021"
 rust-version = "1.57"
 
+[features]
+default = ["rayon"]
+
 [dev-dependencies]
 png = "0.17.7"
 
 [dependencies]
 rgb = "0.8.36"
+rayon = { version = "1.7.0", optional = true }
 
 [package.metadata.docs.rs]
 targets = ["x86_64-unknown-linux-gnu"]