Adds multi-threading and simd resizing via rayon. The old behavior is preserved by disabling the rayon feature which is enabled by default.

* `cargo test`: 1.3s
* `cargo test --no-default-features`: 3.6s
* `cargo test --release`: 0.06s
* `cargo test --no-default-features --release`: 0.1s

Adds 3 new benchmarks for ~normal PNG of RGBA.

| Benchmark Name                  | With "rayon" (ns/iter) | Without "rayon" (ns/iter) | Speedup (ratio)  |
|--------------------------------|------------------------|---------------------------|------------------|
| a_small_rgb                     | 396,614                | 1,413,072                 | 3.56x            |
| a_small_rgba16                  | 371,750                | 1,256,408                 | 3.38x            |
| a_small_rgba16_premultiplied    | 442,669                | 1,788,929                 | 4.04x            |
| huge_stretch                    | 793,616                | 2,462,949                 | 3.10x            |
| init_lanczos                    | 55,823                 | 54,852                    | 0.98x            |
| init_point                      | 24,271                 | 24,211                    | 1.00x            |
| precomputed_large               | 1,497,825              | 5,504,814                 | 3.68x            |
| precomputed_small               | 323,097                | 523,837                   | 1.62x            |
| precomputed_small_16bit         | 353,681                | 786,471                   | 2.22x            |
| recomputed_small                | 343,618                | 538,391                   | 1.57x            |
| tiny                            | 4,539,185              | 34,860,958                | 7.68x            |
| *downsample_rgba                 | 20,384,775             | 149,915,583               | 7.35x            |
| *downsample_tall_rgba            | 13,904,525             | 82,274,604                | 5.92x            |
| *downsample_wide_rgba            | 527,977                | 2,149,650                 | 4.07x            |



<details>

<summary>Benchmarks</summary>

With rayon:
```
running 14 tests
test a_small_rgb                  ... bench:     396,614 ns/iter (+/- 75,382)
test a_small_rgba16               ... bench:     371,750 ns/iter (+/- 53,490)
test a_small_rgba16_premultiplied ... bench:     442,669 ns/iter (+/- 83,742)
test downsample_rgba              ... bench:  20,384,775 ns/iter (+/- 1,149,785)
test downsample_tall_rgba         ... bench:  13,904,525 ns/iter (+/- 2,720,582)
test downsample_wide_rgba         ... bench:     527,977 ns/iter (+/- 126,570)
test huge_stretch                 ... bench:     793,616 ns/iter (+/- 163,863)
test init_lanczos                 ... bench:      55,823 ns/iter (+/- 676)
test init_point                   ... bench:      24,271 ns/iter (+/- 794)
test precomputed_large            ... bench:   1,497,825 ns/iter (+/- 269,346)
test precomputed_small            ... bench:     323,097 ns/iter (+/- 55,199)
test precomputed_small_16bit      ... bench:     353,681 ns/iter (+/- 36,704)
test recomputed_small             ... bench:     343,618 ns/iter (+/- 43,688)
test tiny                         ... bench:   4,539,185 ns/iter (+/- 311,515)
```

With --no-default-features:
```
running 14 tests
test a_small_rgb                  ... bench:   1,413,072 ns/iter (+/- 25,526)
test a_small_rgba16               ... bench:   1,256,408 ns/iter (+/- 41,096)
test a_small_rgba16_premultiplied ... bench:   1,788,929 ns/iter (+/- 9,402)
test downsample_rgba              ... bench: 149,915,583 ns/iter (+/- 4,877,117)
test downsample_tall_rgba         ... bench:  82,274,604 ns/iter (+/- 1,380,165)
test downsample_wide_rgba         ... bench:   2,149,650 ns/iter (+/- 12,188)
test huge_stretch                 ... bench:   2,462,949 ns/iter (+/- 100,948)
test init_lanczos                 ... bench:      54,852 ns/iter (+/- 462)
test init_point                   ... bench:      24,211 ns/iter (+/- 603)
test precomputed_large            ... bench:   5,504,814 ns/iter (+/- 249,495)
test precomputed_small            ... bench:     523,837 ns/iter (+/- 8,511)
test precomputed_small_16bit      ... bench:     786,471 ns/iter (+/- 36,292)
test recomputed_small             ... bench:     538,391 ns/iter (+/- 61,603)
test tiny                         ... bench:  34,860,958 ns/iter (+/- 279,604)
```

</details>


<details>

<summary>SIMD</summary>

I confirmed the hottest loop is getting SIMD'd by the compiler with cargo-show-asm and pulling out the horizontal resampling into a named function temporarily.

```rust
 #[inline(never)]
    pub fn horizontal_resample(&self, tmp_ptr: &AtomicPtr<Format::Accumulator>, pix_fmt: &Format, row: &[Format::InputPixel], y: usize) {
        // Acquire a safe reference to the current position in the temporary buffer.
        let tmp_raw_ptr = tmp_ptr.load(std::sync::atomic::Ordering::Relaxed);

        // For each pixel in the row, calculate the horizontal resampling and store the result.
        for (x, col) in self.scale.coeffs_w.iter().enumerate() {
            let mut accum = Format::new();
            let in_px = &row[col.start..col.start + col.coeffs.len()];
            // hottest loop
            for (coeff, in_px) in col.coeffs.iter().copied().zip(in_px.iter().copied()) {
                pix_fmt.add(&mut accum, in_px, coeff);
            }

            // Determine the location in the temporary buffer to store the result.
            let pixel_offset = y * self.scale.coeffs_w.len() + x;
            // Write the accumulated value to the temporary buffer.
            unsafe {
                *tmp_raw_ptr.add(pixel_offset) = accum;
            }
        }
    }
```

Converting everything to floats before hand (ucvtf) and getting more SIMD in the hot loop was actually slower, likely due to iterating over f32 quads instead of u8 quads. RGB(A) already gets vectorized pretty well. Grayscale does the least from what I could tell.

```
resize::Resizer<Format>::horizontal_resample:
Lfunc_begin153:
            // /Users/jwtrueb/Desktop/workspace/resize/src/lib.rs : 392
                pub fn horizontal_resample(&self, tmp_ptr: &AtomicPtr<Format::Accumulator>, pix_fmt: &Format, row: &[Format::InputPixel], y: usize) {
      ...
LBB153_3:
        add x9, x9, #24
            // /Users/jwtrueb/Desktop/workspace/resize/src/lib.rs : 405
                let pixel_offset = y * self.scale.coeffs_w.len() + x;
        add x14, x11, x12
        add x11, x11, #1
            // /rustc/5680fa18feaa87f3ff04063800aec256c3d4b4be/library/core/src/ptr/mut_ptr.rs : 1020
                unsafe { intrinsics::offset(self, count) }
        madd x14, x14, x13, x8
            // /Users/jwtrueb/Desktop/workspace/resize/src/lib.rs : 408
                *tmp_raw_ptr.add(pixel_offset) = accum;
        str d1, [x14]
        str s2, [x14, #8]
            // /rustc/5680fa18feaa87f3ff04063800aec256c3d4b4be/library/core/src/slice/iter/macros.rs : 141
                if is_empty!(self) {
        cmp x9, x10
        b.eq LBB153_9
LBB153_4:
            // /Users/jwtrueb/Desktop/workspace/resize/src/lib.rs : 399
                let in_px = &row[col.start..col.start + col.coeffs.len()];
        ldp x14, x0, [x9, #8]
        adds x1, x14, x0
            // /rustc/5680fa18feaa87f3ff04063800aec256c3d4b4be/library/core/src/slice/index.rs : 408
                if self.start > self.end {
        b.hs LBB153_10
            // /rustc/5680fa18feaa87f3ff04063800aec256c3d4b4be/library/core/src/slice/index.rs : 410
                } else if self.end > slice.len() {
        cmp x1, x3
        b.hi LBB153_11
            // /rustc/5680fa18feaa87f3ff04063800aec256c3d4b4be/library/core/src/iter/adapters/zip.rs : 270
                if self.index < self.len {
        cbz x14, LBB153_2
        add x15, x0, x0, lsl #1
        add x15, x2, x15
        ldr x16, [x9]
        add x16, x16, #16
        movi d1, #0000000000000000
        movi d2, #0000000000000000
LBB153_8:
            // /rustc/5680fa18feaa87f3ff04063800aec256c3d4b4be/library/core/src/iter/adapters/copied.rs : 103
                *unsafe { try_get_unchecked(&mut self.it, idx) }
        ldr s3, [x16], #4
        ldrb w17, [x15, #2]
        ldrh w0, [x15], #3
        bfi w0, w17, #16, #8
        lsr w1, w0, #8
        uxth w17, w17
        fmov s4, w0
        mov.s v4[1], w1
            // /Users/jwtrueb/Desktop/workspace/resize/src/px.rs : 179
                self as f32
        and.8b v4, v4, v0
        ucvtf.2s v4, v4
            // /Users/jwtrueb/Desktop/workspace/resize/src/px.rs : 39
                acc.r += inp.r.to_float() * coeff;
        fmul.2s v4, v4, v3[0]
        fadd.2s v1, v1, v4
            // /Users/jwtrueb/Desktop/workspace/resize/src/px.rs : 179
                self as f32
        ucvtf s4, w17
            // /Users/jwtrueb/Desktop/workspace/resize/src/px.rs : 41
                acc.b += inp.b.to_float() * coeff;
        fmul s3, s3, s4
        fadd s2, s2, s3
            // /rustc/5680fa18feaa87f3ff04063800aec256c3d4b4be/library/core/src/iter/adapters/zip.rs : 270
                if self.index < self.len {
        subs x14, x14, #1
        b.ne LBB153_8
        b LBB153_3
```
</summary>