@kornelski 2023-08-29
I'm a little concerned that it parallelizes in both dimensions at once. That's millions of tasks vs much fewer cores, and this granularity probably adds overhead. 

Have you tried parallelizing only lines?

Also you may be able to reduce overhead by using https://docs.rs/rayon/latest/rayon/iter/trait.IndexedParallelIterator.html#method.with_min_len to avoid rayon making too-small tasks that are cheaper to execute than to enqueue. Perhaps even switch upfront to non-rayon version for small images.